fix metrics accessors

+++ processor/transformprocessor/internal/metrics/metrics.go
@@ -600,23 +600,23 @@ func accessCount() pathGetSetter {
 		getter: func(ctx common.TransformContext) interface{} {
 			switch ctx.GetItem().(type) {
 			case pmetric.HistogramDataPoint:
-				return ctx.GetItem().(pmetric.HistogramDataPoint).Count()
+				return int64(ctx.GetItem().(pmetric.HistogramDataPoint).Count())
 			case pmetric.ExponentialHistogramDataPoint:
-				return ctx.GetItem().(pmetric.ExponentialHistogramDataPoint).Count()
+				return int64(ctx.GetItem().(pmetric.ExponentialHistogramDataPoint).Count())
 			case pmetric.SummaryDataPoint:
-				return ctx.GetItem().(pmetric.SummaryDataPoint).Count()
+				return int64(ctx.GetItem().(pmetric.SummaryDataPoint).Count())
 			}
 			return nil
 		},
 		setter: func(ctx common.TransformContext, val interface{}) {
-			if newCount, ok := val.(uint64); ok {
+			if newCount, ok := val.(int64); ok {
 				switch ctx.GetItem().(type) {
 				case pmetric.HistogramDataPoint:
-					ctx.GetItem().(pmetric.HistogramDataPoint).SetCount(newCount)
+					ctx.GetItem().(pmetric.HistogramDataPoint).SetCount(uint64(newCount))
 				case pmetric.ExponentialHistogramDataPoint:
-					ctx.GetItem().(pmetric.ExponentialHistogramDataPoint).SetCount(newCount)
+					ctx.GetItem().(pmetric.ExponentialHistogramDataPoint).SetCount(uint64(newCount))
 				case pmetric.SummaryDataPoint:
-					ctx.GetItem().(pmetric.SummaryDataPoint).SetCount(newCount)
+					ctx.GetItem().(pmetric.SummaryDataPoint).SetCount(uint64(newCount))
 				}
 			}
 		},
@@ -696,15 +696,15 @@ func accessScale() pathGetSetter {
 		getter: func(ctx common.TransformContext) interface{} {
 			switch ctx.GetItem().(type) {
 			case pmetric.ExponentialHistogramDataPoint:
-				return ctx.GetItem().(pmetric.ExponentialHistogramDataPoint).Scale()
+				return int64(ctx.GetItem().(pmetric.ExponentialHistogramDataPoint).Scale())
 			}
 			return nil
 		},
 		setter: func(ctx common.TransformContext, val interface{}) {
-			if newScale, ok := val.(int32); ok {
+			if newScale, ok := val.(int64); ok {
 				switch ctx.GetItem().(type) {
 				case pmetric.ExponentialHistogramDataPoint:
-					ctx.GetItem().(pmetric.ExponentialHistogramDataPoint).SetScale(newScale)
+					ctx.GetItem().(pmetric.ExponentialHistogramDataPoint).SetScale(int32(newScale))
 				}
 			}
 		},
@@ -716,15 +716,15 @@ func accessZeroCount() pathGetSetter {
 		getter: func(ctx common.TransformContext) interface{} {
 			switch ctx.GetItem().(type) {
 			case pmetric.ExponentialHistogramDataPoint:
-				return ctx.GetItem().(pmetric.ExponentialHistogramDataPoint).ZeroCount()
+				return int64(ctx.GetItem().(pmetric.ExponentialHistogramDataPoint).ZeroCount())
 			}
 			return nil
 		},
 		setter: func(ctx common.TransformContext, val interface{}) {
-			if newZeroCount, ok := val.(uint64); ok {
+			if newZeroCount, ok := val.(int64); ok {
 				switch ctx.GetItem().(type) {
 				case pmetric.ExponentialHistogramDataPoint:
-					ctx.GetItem().(pmetric.ExponentialHistogramDataPoint).SetZeroCount(newZeroCount)
+					ctx.GetItem().(pmetric.ExponentialHistogramDataPoint).SetZeroCount(uint64(newZeroCount))
 				}
 			}
 		},
@@ -756,15 +756,15 @@ func accessPositiveOffset() pathGetSetter {
 		getter: func(ctx common.TransformContext) interface{} {
 			switch ctx.GetItem().(type) {
 			case pmetric.ExponentialHistogramDataPoint:
-				return ctx.GetItem().(pmetric.ExponentialHistogramDataPoint).Positive().Offset()
+				return int64(ctx.GetItem().(pmetric.ExponentialHistogramDataPoint).Positive().Offset())
 			}
 			return nil
 		},
 		setter: func(ctx common.TransformContext, val interface{}) {
-			if newPositiveOffset, ok := val.(int32); ok {
+			if newPositiveOffset, ok := val.(int64); ok {
 				switch ctx.GetItem().(type) {
 				case pmetric.ExponentialHistogramDataPoint:
-					ctx.GetItem().(pmetric.ExponentialHistogramDataPoint).Positive().SetOffset(newPositiveOffset)
+					ctx.GetItem().(pmetric.ExponentialHistogramDataPoint).Positive().SetOffset(int32(newPositiveOffset))
 				}
 			}
 		},
@@ -816,15 +816,15 @@ func accessNegativeOffset() pathGetSetter {
 		getter: func(ctx common.TransformContext) interface{} {
 			switch ctx.GetItem().(type) {
 			case pmetric.ExponentialHistogramDataPoint:
-				return ctx.GetItem().(pmetric.ExponentialHistogramDataPoint).Negative().Offset()
+				return int64(ctx.GetItem().(pmetric.ExponentialHistogramDataPoint).Negative().Offset())
 			}
 			return nil
 		},
 		setter: func(ctx common.TransformContext, val interface{}) {
-			if newNegativeOffset, ok := val.(int32); ok {
+			if newNegativeOffset, ok := val.(int64); ok {
 				switch ctx.GetItem().(type) {
 				case pmetric.ExponentialHistogramDataPoint:
-					ctx.GetItem().(pmetric.ExponentialHistogramDataPoint).Negative().SetOffset(newNegativeOffset)
+					ctx.GetItem().(pmetric.ExponentialHistogramDataPoint).Negative().SetOffset(int32(newNegativeOffset))
 				}
 			}
 		},
